Отделил модель робота (изменение координат и направление робота) от его отрисовки
