+++ src/main/java/org/iffomko/gui/GameVisualizer.java
@@ -1,7 +1,6 @@
 package org.iffomko.gui;
 
-import org.iffomko.savers.Savable;
-
+import org.iffomko.robot.Robot;
 import java.awt.Color;
 import java.awt.EventQueue;
 import java.awt.Graphics;
@@ -10,7 +9,6 @@ import java.awt.Point;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.awt.geom.AffineTransform;
-import java.util.Map;
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -22,6 +20,7 @@ import javax.swing.JPanel;
 public class GameVisualizer extends JPanel
 {
     private final Timer m_timer = initTimer();
+    private final Robot robot;
 
     /**
      * Создается таймер, который называется генератор событий
@@ -33,19 +32,12 @@ public class GameVisualizer extends JPanel
         return timer;
     }
 
-    private volatile double m_robotPositionX = 100;
-    private volatile double m_robotPositionY = 100;
-    private volatile double m_robotDirection = 0;
-
     private volatile int m_targetPositionX = 150;
     private volatile int m_targetPositionY = 100;
 
-    private static final double maxVelocity = 0.1;
-    private static final double maxAngularVelocity = 0.001;
-
     public GameVisualizer()
     {
-
+        robot = new Robot();
 
         m_timer.schedule(new TimerTask()
         {
@@ -94,122 +86,11 @@ public class GameVisualizer extends JPanel
     }
 
     /**
-     * Высчитывает расстояние между двумя точками (x1, y1) и (x2, y2)
-     * @param x1 - координата x первой точки
-     * @param y1 - координата y первой точки
-     * @param x2 - координата x второй точки
-     * @param y2 - координата y второй точки
-     * @return - расстояние
-     */
-    private static double distance(double x1, double y1, double x2, double y2)
-    {
-        double diffX = x1 - x2;
-        double diffY = y1 - y2;
-        return Math.sqrt(diffX * diffX + diffY * diffY);
-    }
-
-    /**
-     * Вычисляет тета-угол между двумя точками
-     * @param fromX - координата по оси OX первой точки
-     * @param fromY - координата по оси OY первой точки
-     * @param toX - координата по оси OX второй точки
-     * @param toY - координата по оси OY второй точки
-     * @return - тета-угол
-     */
-    private static double angleTo(double fromX, double fromY, double toX, double toY)
-    {
-        double diffX = toX - fromX;
-        double diffY = toY - fromY;
-
-        return asNormalizedRadians(Math.atan2(diffY, diffX));
-    }
-
-    /**
      * Метод, который вызывается каждый раз для обновления координат моделей
      */
     protected void onModelUpdateEvent()
     {
-        double distance = distance(m_targetPositionX, m_targetPositionY,
-                m_robotPositionX, m_robotPositionY); // расстояние между точкой назначения
-        if (distance < 0.5) {
-            return;
-        }
-
-        double velocity = maxVelocity;
-        double angleToTarget = angleTo(m_robotPositionX, m_robotPositionY, m_targetPositionX, m_targetPositionY);
-        double angularVelocity = 0;
-
-        if (angleToTarget > m_robotDirection) {
-            angularVelocity = maxAngularVelocity;
-        } else if (angleToTarget < m_robotDirection)
-        {
-            angularVelocity = -maxAngularVelocity;
-        }
-
-        moveRobot(velocity, angularVelocity, 10);
-    }
-
-    /**
-     * Приводит входное значение <code>value</code> к нормализованным значениям, т. е. чтобы не вышло за границы <code>[min, max]</code>
-     * @param value - значение, которое нужно привести
-     * @param min - минимальное значение, которое может принимать <code>value</code>
-     * @param max - максимальное значение, которое может принимать <code>value</code>
-     * @return - приведенное значение
-     */
-    private static double applyLimits(double value, double min, double max)
-    {
-        if (value < min)
-            return min;
-        if (value > max)
-            return max;
-        return value;
-    }
-
-    /**
-     * Передвигает робота
-     * @param velocity - линейная скорость передвижения робота
-     * @param angularVelocity - угловая скорость передвижения робота
-     * @param duration - промежуток, в течении которого он двигается
-     */
-    private void moveRobot(double velocity, double angularVelocity, double duration)
-    {
-        velocity = applyLimits(velocity, 0, maxVelocity); // нормализуем обычную скорость робота
-        angularVelocity = applyLimits(angularVelocity, -maxAngularVelocity, maxAngularVelocity); // Нормализуем угловую скорость робота, т. е. скорость при вращении
-        double newX = m_robotPositionX + velocity *
-                (Math.sin(m_robotDirection  + angularVelocity * duration) - Math.sin(m_robotDirection)) / angularVelocity;
-        if (!Double.isFinite(newX))
-        {
-            newX = m_robotPositionX + velocity * duration * Math.cos(m_robotDirection);
-        }
-        double newY = m_robotPositionY - velocity  *
-                (Math.cos(m_robotDirection  + angularVelocity * duration) -
-                        Math.cos(m_robotDirection)) / angularVelocity;
-        if (!Double.isFinite(newY))
-        {
-            newY = m_robotPositionY + velocity * duration * Math.sin(m_robotDirection);
-        }
-        m_robotPositionX = newX;
-        m_robotPositionY = newY;
-        double newDirection = asNormalizedRadians(m_robotDirection + angularVelocity * duration); // меняем направление робота
-        m_robotDirection = newDirection;
-    }
-
-    /**
-     * Делает угол положительным (0 <= angle < 2*Math.PI)
-     * @param angle - входящий угол в радианах
-     * @return - нормализованный угол
-     */
-    private static double asNormalizedRadians(double angle)
-    {
-        while (angle < 0)
-        {
-            angle += 2*Math.PI;
-        }
-        while (angle >= 2*Math.PI)
-        {
-            angle -= 2*Math.PI;
-        }
-        return angle;
+        robot.move(m_targetPositionX, m_targetPositionY, 10);
     }
 
     /**
@@ -222,6 +103,7 @@ public class GameVisualizer extends JPanel
         return (int)(value + 0.5);
     }
 
+
     /**
      * Отрисовывает графику игры
      * @param g -  <code>Graphics</code> контекст, в котором нужно отрисовывать игру
@@ -231,7 +113,7 @@ public class GameVisualizer extends JPanel
     {
         super.paint(g);
         Graphics2D g2d = (Graphics2D)g;
-        drawRobot(g2d, round(m_robotPositionX), round(m_robotPositionY), m_robotDirection);
+        drawRobot(g2d, round(robot.getX()), round(robot.getY()), robot.getDirection());
         drawTarget(g2d, m_targetPositionX, m_targetPositionY);
     }
 
@@ -270,8 +152,8 @@ public class GameVisualizer extends JPanel
      */
     private void drawRobot(Graphics2D g, int x, int y, double direction)
     {
-        int robotCenterX = round(m_robotPositionX);
-        int robotCenterY = round(m_robotPositionY);
+        int robotCenterX = round(robot.getX());
+        int robotCenterY = round(robot.getY());
         AffineTransform t = AffineTransform.getRotateInstance(direction, robotCenterX, robotCenterY); // поворачивает
         // робота на угол direction относительно координат robotCenterX, robotCenterY
